Users ajout JWT + debug connexion et inscription

+++ projetTut2BackEnd/src/Controller/Users/UserController.php
@@ -4,6 +4,7 @@ namespace App\Controller\Users;
 
 use App\Controller\AbstractController;
 use App\Entity\Game;
+use App\Manager\TokenManager;
 use PDO;
 use Symfony\Component\HttpFoundation\JsonResponse;
 use Symfony\Component\HttpFoundation\Request;
@@ -34,14 +35,29 @@ class UserController extends AbstractController
 
         $searchParams = $this->parseRequestContent($requestContent);
 
-        $req = $this->bdd->prepare("INSERT INTO users_games VALUES (:user, :game)");
+        $data = $this->getToken($request);
 
-        $req->bindParam(':user', $searchParams['user'], PDO::PARAM_INT);
-        $req->bindParam(':game', $searchParams['game'], PDO::PARAM_INT);
+        if ($this->checkToken($data, $searchParams['user'])) {
+            $select = $this->bdd->prepare("SELECT * FROM users_games WHERE user = :user AND game = :game");
 
-        if ($req->execute()) {
-            return new Response(true);
+            $select->bindParam(':user', $searchParams['user'], PDO::PARAM_INT);
+            $select->bindParam(':game', $searchParams['game'], PDO::PARAM_INT);
+
+            $select->execute();
+
+            if ($select->rowCount() === 0) { //On vérifie si la relation n'existe pas déjà
+                $req = $this->bdd->prepare("INSERT INTO users_games VALUES (:user, :game)");
+
+                $req->bindParam(':user', $searchParams['user'], PDO::PARAM_INT);
+                $req->bindParam(':game', $searchParams['game'], PDO::PARAM_INT);
+
+                if ($req->execute()) {
+                    return new Response(true);
+                }
+            }
+            return new Response(false);
         }
+        return new Response(false);
     }
 
     /**
@@ -55,14 +71,20 @@ class UserController extends AbstractController
 
         $searchParams = $this->parseRequestContent($requestContent);
 
-        $req = $this->bdd->prepare("DELETE FROM users_games WHERE user = :user AND game = :game");
+        $data = $this->getToken($request);
 
-        $req->bindParam(':user', $searchParams['user'], PDO::PARAM_INT);
-        $req->bindParam(':game', $searchParams['game'], PDO::PARAM_INT);
+        if ($this->checkToken($data, $searchParams['user'])) {
+            $req = $this->bdd->prepare("DELETE FROM users_games WHERE user = :user AND game = :game");
 
-        if ($req->execute()) {
-            return new Response(true);
+            $req->bindParam(':user', $searchParams['user'], PDO::PARAM_INT);
+            $req->bindParam(':game', $searchParams['game'], PDO::PARAM_INT);
+
+            if ($req->execute()) {
+                return new Response(true);
+            }
+            return new Response(false);
         }
+        return new Response(false);
     }
 
     /**
@@ -77,62 +99,85 @@ class UserController extends AbstractController
 
         $searchParams = $this->parseRequestContent($requestContent);
 
-        $req = $this->bdd->prepare('SELECT game FROM users_games WHERE user = :user');
+        $data = $this->getToken($request);
 
-        $req->execute(array(
-            'user' => $searchParams['user']
-        ));
+        if ($this->checkToken($data, $searchParams['user'])) {
+            $req = $this->bdd->prepare('SELECT game FROM users_games WHERE user = :user');
 
-        $resultSQL = $req->fetchAll();
+            $req->execute(array(
+                'user' => $searchParams['user']
+            ));
 
-        $gamesByPage = 8;
-        $page = $searchParams['page'];
-        if ($page < 1) {
-            $page = 1;
-        }
-        $params = [
-            'index' => 'steam',
-            'size' => $gamesByPage,
-            'from' => ($page - 1) * $gamesByPage,
-            'body' => [
-                'query' => [
-                    'bool' => [
-                        'should' => [],
+            $resultSQL = $req->fetchAll();
+
+            $gamesByPage = 8;
+            $page = $searchParams['page'];
+            if ($page < 1) {
+                $page = 1;
+            }
+            $params = [
+                'index' => 'steam',
+                'size' => $gamesByPage,
+                'from' => ($page - 1) * $gamesByPage,
+                'body' => [
+                    'query' => [
+                        'bool' => [
+                            'should' => [],
+                        ],
                     ],
                 ],
-            ],
-        ];
+            ];
 
-        $queryParams = [];
+            $queryParams = [];
 
-        foreach ($resultSQL as $game) {
-            array_push($queryParams, array("terms" => array('data.appid' =>  (array)$game[0])));
+            foreach ($resultSQL as $game) {
+                array_push($queryParams, array("terms" => array('data.appid' =>  (array)$game[0])));
+            }
+
+            $params['body']['query']['bool']['should'] = $queryParams;
+
+            $result = $this->client->search($params);
+
+            $games = ['games' => []];
+            foreach ($result['hits']['hits'] as $gameInfos) {
+                $idgame = $gameInfos['_source']['data']['appid'];
+
+                $image = $this->createImage($idgame);
+
+                $game = new Game();
+                $game->hydrate($gameInfos['_source']['data']);
+                $game->setImage($image);
+                $game->setId($gameInfos['_id']);
+                array_push($games['games'], json_decode($this->serializer->serialize($game, 'json')));
+            }
+
+            unset($params['size']);
+            unset($params['page']);
+            unset($params['from']);
+
+            $totalGames = $this->client->count($params);
+
+            $games['nbPages'] = ceil($totalGames['count'] / $gamesByPage);
+            return new JsonResponse($games);
         }
+        return new Response(false);
+    }
 
-        $params['body']['query']['bool']['should'] = $queryParams;
+    public function getToken(Request $request)
+    {
+        $authorizationHeader = $request->headers->get('Authorization');
+        $authorizationHeaderArray = explode(' ', $authorizationHeader);
+        $token = $authorizationHeaderArray[0] ?? null;
+        $data = (new TokenManager())->decode($token);
 
-        $result = $this->client->search($params);
+        return $data;
+    }
 
-        $games = ['games' => []];
-        foreach ($result['hits']['hits'] as $gameInfos) {
-            $idgame = $gameInfos['_source']['data']['appid'];
-
-            $image = $this->createImage($idgame);
-
-            $game = new Game();
-            $game->hydrate($gameInfos['_source']['data']);
-            $game->setImage($image);
-            $game->setId($gameInfos['_id']);
-            array_push($games['games'], json_decode($this->serializer->serialize($game, 'json')));
+    public function checkToken($data, $id)
+    {
+        if ($data['exp'] > time() && $data['id'] === $id) {
+            return true;
         }
-
-        unset($params['size']);
-        unset($params['page']);
-        unset($params['from']);
-
-        $totalGames = $this->client->count($params);
-
-        $games['nbPages'] = ceil($totalGames['count'] / $gamesByPage);
-        return new JsonResponse($games);
+        return false;
     }
 }
